Theme JSON: include block style variations in path only output of get_block_nodes
An `$include_node_paths_only` option to `get_block_nodes()` was introduced to improve performance.

When set to `true`, this option tells the function to only return paths, and not selectors, for consumers that only needed paths to style values.

For one of the conditional blocks, block style variations wasn't included.

This commit adds them to the array of paths following the existing model `$node[]['path' => [], 'variations' => ['path' => []]]`.

Follow-up to [61858].

Props aaronrobertshaw, ramonopoly.
Fixes #62399.




git-svn-id: https://develop.svn.wordpress.org/trunk@59418 602fd350-edb4-49c9-b593-d223f7449a82