Toolbar: Allow the Learn WordPress link to be localized.
The Learn WordPress website is supposed to automatically redirect to the correct locale according to the browser's language settings, however that may not work as expected in some cases.

This commit brings consistency with the other WordPress.org links, which can be localized as appropriate.

Follow-up to [56720].

Props timse201, ruturajraval2305, yogeshbhutkar, ajayghaghretiya-multidots, swissspidy, sabernhardt, im3dabasia1, mukesh27.
Fixes #62459.

git-svn-id: https://develop.svn.wordpress.org/trunk@59425 602fd350-edb4-49c9-b593-d223f7449a82