Build/Test Tools: Add `repository` input to support JSON reading workflow.
`actions/checkout` will always checkout the current repository unless the `repository` input is specified. This updates the `reusable-support-json-reader-v1.yml` workflow to always default to reading the JSON files from `wordpress-develop`.

A `repository` has also been added to the workflow to allow a different set of JSON files to be read if desired.

See #62221.

git-svn-id: https://develop.svn.wordpress.org/trunk@59483 602fd350-edb4-49c9-b593-d223f7449a82