Administration: Remove useless arrow icon from WordPress admin menu.
This changeset deletes the arrow that is typically added next to WordPress admin menu items that have submenus. The `.wp-menu-arrow` element is no longer visible since the WP 3.8 redesign, but the HTML and CSS remained. With this changeset, the HTML generating the arrow is removed, and the corresponding CSS styling is deleted.

Props helen, azaozz, jbkkd, pbearne, flixos90.
Fixes #26960.




git-svn-id: https://develop.svn.wordpress.org/trunk@59690 602fd350-edb4-49c9-b593-d223f7449a82