Build/Test Tools: Configure UglifyJS to preserve previous behavior.
As of UglifyJS >= 3.18.0, the default behavior is to process input as an ES module. This updates the relevant configurations to ensure the build process continues to use the previous behavior to avoid JavaScript errors in the minified versions of files.

Follow up to [58563], [58586], and [59509].

Props siliconforks, nataliat2004, poena, mai21, SergeyBiryukov.
Fixes #62767. See #61519, #62220.

git-svn-id: https://develop.svn.wordpress.org/trunk@59768 602fd350-edb4-49c9-b593-d223f7449a82